notes_backend/src/api/database.py
import os
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker
from dotenv import load_dotenv
from .models import Base

load_dotenv()

# Use .env for database URL; default to SQLite
DATABASE_URL = os.getenv("DATABASE_URL", "sqlite:///./notes.db")

connect_args = {"check_same_thread": False} if DATABASE_URL.startswith("sqlite") else {}

engine = create_engine(
    DATABASE_URL, connect_args=connect_args
)

SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

# PUBLIC_INTERFACE
def init_db():
    """Initialize the database and create tables."""
    Base.metadata.create_all(bind=engine)


# PUBLIC_INTERFACE
def get_db():
    """FastAPI dependency for DB session."""
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()
